show rested points as uInteger value, and not as float value, what looks bad..

+++ src/server/game/Entities/Player/Player.cpp
@@ -11327,7 +11327,7 @@ uint32 Player::GetXPRestBonus(uint32 xp)
 
     SetRestBonus(GetRestBonus() - rested_bonus);
 
-    TC_LOG_INFO("entities.player", "Player gain %u xp (+ %u Rested Bonus). Rested points=%f", xp+rested_bonus, rested_bonus, GetRestBonus());
+    TC_LOG_INFO("entities.player", "Player gain %u xp (+ %u Rested Bonus). Rested points=%u", xp + rested_bonus, rested_bonus, (uint32)GetRestBonus());
     return rested_bonus;
 }
